move beacon to the Engine directory
