+++ inc/Engine/Admin/ServiceProvider.php
@@ -24,7 +24,6 @@ class ServiceProvider extends AbstractServiceProvider {
 		'settings_page_subscriber',
 		'deactivation_intent_render',
 		'deactivation_intent_subscriber',
-		'beacon_subscriber',
 		'hummingbird_subscriber',
 	];
 
@@ -45,8 +44,6 @@ class ServiceProvider extends AbstractServiceProvider {
 			->withArgument( $this->getContainer()->get( 'deactivation_intent_render' ) )
 			->withArgument( $this->getContainer()->get( 'options_api' ) )
 			->withArgument( $this->getContainer()->get( 'options' ) );
-		$this->getContainer()->share( 'beacon_subscriber', 'WP_Rocket\Subscriber\Admin\Settings\Beacon_Subscriber' )
-			->withArgument( $this->getContainer()->get( 'beacon' ) );
 		$this->getContainer()->share( 'hummingbird_subscriber', 'WP_Rocket\Subscriber\Third_Party\Plugins\Optimization\Hummingbird_Subscriber' )
 			->withArgument( $this->getContainer()->get( 'options' ) );
 	}
